fix: move validate_google_service_account to service_account.py
Resolves LLM checks CI failure: validation.py drops from 594 to 497 lines
(under the 500-line limit). The function is self-contained and logically
belongs in service_account.py alongside other service-account functions.

Uses lazy import pattern to avoid circular dependency.
